Changes to Enable SSH for Repository hosting
We use Git over SSH for repository hosting. Adding the ability for
this cookbook to be used for configuring the phab host for this.
Runs an additional SSH Daemon on port 617 (configurable) for this
purpose.

Test cases added to ensure the daemon is installed and configured.

N.B. This is only for Ubuntu 14.04 or newer, due to need for
AuthorizedKeysCommand which came in OpenSSH 6.2.

Functionality is disabled on Ubuntu 12.04, and test cases are skipped
in that case.