Fix service startup and deployment
Firstly with mysql-server, it was being restarted with phab config
delayed, which during initial run (or during kitchen test run), meant
that phd service was being created before mysql had been restarted.
Thus when mysql was restarted it was restarting phd, causing the
behaviour noticed previously in a comment.

Secondly, with the `run_storage_upgrade` task, switched to using a guard
`not_if` rather than the preivous mechanism, as this should be more robust
and is cleaner.

Lastly with the phd upstart script, configured it to not depend on mysql starting
when mysql server is not local.